pyproject.toml
[project]
name = "queryscope"
version = "0.1.0"
description = "Django ORM profiling dashboard"
readme = "README.md"
requires-python = ">=3.12"

dependencies = [
    "django>=5.1",
    "djangorestframework>=3.15",
    "psycopg2-binary>=2.9",
    "django-debug-toolbar>=4.0",
    "django-silk>=5.0",
    "factory-boy>=3.3",
    "faker>=24.0",
    "django-redis>=5.4",
    "python-decouple>=3.8",
]

[dependency-groups]
dev = [
    "playwright>=1.58.0",
    "pytest>=8.0",
    "pytest-django>=4.8",
    "pytest-factoryboy>=2.6",
]

[tool.uv]
package = false

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "queryscope.settings.dev"
python_files = ["tests/test_*.py"]
